Introduced aspects: Like types, but different.
Like types, aspects can be used as a taxonomy for classification, and
like types, aspects can provide more context about properties (eg. an
integer value can actually be an enum).

Unlike types, a Propertyset can have more than one type.  It is not
multiple inheritance, when an aspect is used to "view" a Propertyset,
its type is added. If a particular value name set in this context also
exist in a different aspect, the last set value will overwrite the
previous setting.