Inject a PropertysetManager service and do simple operations on it in a test.
The test is run as a regular JUnit test, but will start its own Karaf
runime and test the modelstore bundle there.

Note that while the test itself will run from inside the JUnit test view
in eclipse, the test requires the actual .jar file of the bundle to be
able to run, and m2e will not build the actual jar file.

So until the modelstore.implementation project has been actually built
with maven from the command line, the Pax Exam tests will fail.