Added maven module for running the Apache Felix gogo shell.
The added module doesn't build any distributable artifacts during "mvn
install".  The module is to make it easy to start a gogo shell with the
bundles built by the modelstore project.

To start the shell:
 cd modelstore.gogoshell
 mvn install pax:provision

After the startup the bundles can be examined.  The startup and shutdown
of the shell are also of interest for the bundles' OSGi health: there
should be no error messages with stack traces in the shell startup and
shutdown.