Store lastmodifiedtimes for propertysets.
The lastmodifiedtimes are stored in a metadata propertyset that is
serialized first when the context is persisted as JSON.

The PropertysetContext is wrapped in the delegating object
PropertysetContextRecordingMetadata.  The
PropertysetContextRecordingMetadata object will wrap returned
Propertyset objects in a delegating object
PropertysetRecordingSaveTimeTest also implementing Propertyset.  All
setters of PropertysetRecordingSaveTimeTest will set the
lastmodifiedtime of the Propertyset being wrapped.